Moving code around for Dol Patches
