+++ test/games/prime/test_dol_patcher.py
@@ -1,24 +1,16 @@
 from pathlib import Path
-from unittest.mock import patch, MagicMock, ANY
+from unittest.mock import patch, MagicMock
 
 from randovania.games.prime import dol_patcher
-from randovania.interface_common.echoes_user_preferences import EchoesUserPreferences
 
 
-@patch("randovania.games.prime.dol_patcher._apply_beam_cost_patch", autospec=True)
-@patch("randovania.games.prime.dol_patcher._apply_energy_tank_capacity_patch", autospec=True)
-@patch("randovania.games.prime.dol_patcher._apply_game_options_patch", autospec=True)
-@patch("randovania.games.prime.dol_patcher._apply_string_display_patch", autospec=True)
 @patch("randovania.games.prime.dol_patcher.DolFile")
 @patch("randovania.games.prime.dol_patcher._get_dol_path", autospec=True)
 @patch("randovania.games.prime.dol_patcher.find_version_for_dol", autospec=True)
 def test_apply_patches(mock_find_version_for_dol: MagicMock,
                        mock_get_dol_path: MagicMock,
                        mock_dol_file_constructor: MagicMock,
-                       mock_apply_string_display_patch: MagicMock,
-                       mock_apply_game_options_patch: MagicMock,
-                       mock_apply_energy_tank_capacity_patch: MagicMock,
-                       mock_apply_beam_cost_patch: MagicMock,
+                       mocker,
                        ):
     # Setup
     game_root = MagicMock()
@@ -28,6 +20,14 @@ def test_apply_patches(mock_find_version_for_dol: MagicMock,
     mock_find_version_for_dol.return_value = version_patches
     dol_file = mock_dol_file_constructor.return_value
 
+    mock_apply_string = mocker.patch("randovania.games.prime.all_prime_dol_patches.apply_string_display_patch", autospec=True)
+    mock_apply_game_options: MagicMock = mocker.patch(
+        "randovania.games.prime.echoes_dol_patches.apply_game_options_patch", autospec=True)
+    mock_apply_capacity: MagicMock = mocker.patch(
+        "randovania.games.prime.all_prime_dol_patches.apply_energy_tank_capacity_patch", autospec=True)
+    mock_apply_beam_cost_patch: MagicMock = mocker.patch(
+        "randovania.games.prime.echoes_dol_patches.apply_beam_cost_patch", autospec=True)
+
     # Run
     dol_patcher.apply_patches(game_root, game_patches.game_specific, user_preferences)
 
@@ -35,12 +35,12 @@ def test_apply_patches(mock_find_version_for_dol: MagicMock,
     mock_find_version_for_dol.assert_called_once_with(dol_file, dol_patcher.ALL_VERSIONS_PATCHES)
     mock_get_dol_path.assert_called_once_with(game_root)
     mock_dol_file_constructor.assert_called_once_with(mock_get_dol_path.return_value)
-    mock_apply_string_display_patch.assert_called_once_with(version_patches.string_display, dol_file)
-    mock_apply_game_options_patch.assert_called_once_with(
+    mock_apply_string.assert_called_once_with(version_patches.string_display, dol_file)
+    mock_apply_game_options.assert_called_once_with(
         version_patches.game_options_constructor_address,
         user_preferences, dol_file
     )
-    mock_apply_energy_tank_capacity_patch.assert_called_once_with(
+    mock_apply_capacity.assert_called_once_with(
         version_patches.health_capacity,
         game_patches.game_specific, dol_file
     )
@@ -53,25 +53,3 @@ def test_apply_patches(mock_find_version_for_dol: MagicMock,
 def test_get_dol_path():
     assert dol_patcher._get_dol_path(Path("foo")) == Path("foo", "sys", "main.dol")
 
-
-def test_apply_string_display_patch():
-    offsets = dol_patcher.StringDisplayPatchAddresses(0x15, 0, 0, 0, 0, 0)
-    dol_file = MagicMock()
-
-    # Run
-    dol_patcher._apply_string_display_patch(offsets, dol_file)
-
-    # Assert
-    dol_file.write.assert_called_once_with(0x15, ANY)
-
-
-def test_apply_game_options_patch():
-    offset = 0x15
-    dol_file = MagicMock()
-    user_preferences = EchoesUserPreferences()
-
-    # Run
-    dol_patcher._apply_game_options_patch(offset, user_preferences, dol_file)
-
-    # Assert
-    dol_file.write.assert_called_once_with(53, ANY)
