+++ test/games/prime/test_prime_binary_decoder.py
@@ -27,7 +27,8 @@ def test_simple_round_trip():
         "game_specific": {
             "energy_per_tank": 100.0,
             "safe_zone_heal_per_second": 1.0,
-            "beam_configurations": []
+            "beam_configurations": [],
+            "dangerous_energy_tank": True,
         },
         "starting_location": {
             "world_asset_id": 1006255871,
